put render pipeline cache into a separate file
